Refactor access to memory: from O(n) to O(1)

+++ src/main/java/com/github/jagarsoft/Memory.java
@@ -3,5 +3,5 @@ package com.github.jagarsoft;
 interface Memory {
     void poke(int addr, byte data);
     byte peek(int addr);
-    int getSize();
+    short getSize();
 }
