+++ src/main/java/com/github/jagarsoft/RAMMemory.java
@@ -2,16 +2,16 @@ package com.github.jagarsoft;
 
 public class RAMMemory implements Memory {
         private byte[] ram;
-        public int size;
+        public short size;
 
         public RAMMemory(int size){
             this.ram = new byte[size];
-            this.size = size;
+            this.size = (short)size;
         }
 
         public void poke(int addr, byte data) { ram[addr] = data; }
 
-        public int getSize(){ return size; }
+        public short getSize(){ return size; }
 
         public byte peek(int addr) { return ram[addr]; }
 }
